+++ .fossa.yml
@@ -1173,9 +1173,6 @@ targets:
       target: ':instrumentation:spring:spring-jms:spring-jms-6.0:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:spring:spring-web:spring-web-3.1:javaagent'
-    - type: gradle
-      path: ./
       target: ':instrumentation:spring:spring-web:spring-web-3.1:library'
     - type: gradle
       path: ./
