+++ instrumentation/spring/spring-web/spring-web-3.1/testing/build.gradle.kts
@@ -1,17 +1,5 @@
 plugins {
-  id("otel.javaagent-instrumentation")
-  id("otel.nullaway-conventions")
-}
-
-muzzle {
-  pass {
-    group.set("org.springframework")
-    module.set("spring-web")
-    versions.set("[3.1.0.RELEASE,6)")
-    // these versions depend on javax.faces:jsf-api:1.1 which was released as pom only
-    skip("1.2.1", "1.2.2", "1.2.3", "1.2.4")
-    assertInverse.set(true)
-  }
+  id("otel.javaagent-testing")
 }
 
 dependencies {
