+++ instrumentation/spring/spring-web/spring-web-3.1/testing/src/test/java/io/opentelemetry/javaagent/instrumentation/springweb/v3_1/SpringRestTemplateTest.java
