+++ instrumentation/spring/spring-webmvc/spring-webmvc-3.1/javaagent/build.gradle.kts
@@ -30,7 +30,6 @@ dependencies {
   testInstrumentation(project(":instrumentation:spring:spring-webmvc:spring-webmvc-6.0:javaagent"))
   testInstrumentation(project(":instrumentation:servlet:servlet-3.0:javaagent"))
   testInstrumentation(project(":instrumentation:tomcat:tomcat-7.0:javaagent"))
-  testInstrumentation(project(":instrumentation:spring:spring-web:spring-web-3.1:javaagent"))
 
   testImplementation(project(":instrumentation:spring:spring-webmvc:spring-webmvc-common:testing"))
   testImplementation("com.google.guava:guava")
