+++ instrumentation/spring/spring-webmvc/spring-webmvc-6.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/webmvc/v6_0/springweb/WebApplicationContextInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.springweb.v3_1;
+package io.opentelemetry.javaagent.instrumentation.spring.webmvc.v6_0.springweb;
 
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.extendsClass;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.hasClassesNamed;
@@ -57,9 +57,8 @@ class WebApplicationContextInstrumentation implements TypeInstrumentation {
 
     @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
     public static void onEnter(@Advice.Argument(0) ConfigurableListableBeanFactory beanFactory) {
-      if (beanFactory instanceof BeanDefinitionRegistry
+      if (beanFactory instanceof BeanDefinitionRegistry beanDefinitionRegistry
           && !beanFactory.containsBean("otelAutoDispatcherFilter")) {
-
         // Explicitly loading classes allows to catch any class-loading issue or deal with cases
         // where the class is not visible.
         try {
@@ -77,13 +76,12 @@ class WebApplicationContextInstrumentation implements TypeInstrumentation {
               dispatcherServletClass
                   .getClassLoader()
                   .loadClass(
-                      "org.springframework.web.servlet.v3_1.OpenTelemetryHandlerMappingFilter");
+                      "org.springframework.web.servlet.v6_0.OpenTelemetryHandlerMappingFilter");
           GenericBeanDefinition beanDefinition = new GenericBeanDefinition();
           beanDefinition.setScope(SCOPE_SINGLETON);
           beanDefinition.setBeanClass(clazz);
 
-          ((BeanDefinitionRegistry) beanFactory)
-              .registerBeanDefinition("otelAutoDispatcherFilter", beanDefinition);
+          beanDefinitionRegistry.registerBeanDefinition("otelAutoDispatcherFilter", beanDefinition);
         } catch (ClassNotFoundException ignored) {
           // Ignore
         }
